backend/core/requirements.txt
absl-py==2.3.1
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.13.4
bleach==6.3.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
chromadb==1.1.1
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==3.9.0
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
fastapi==0.116.1
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.6
google-api-core==1.34.1
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.70.0
gql==4.0.0
graphql-core==3.2.7
greenlet==3.2.3
griffe==1.9.0
groq==1.1.1
grpcio==1.74.0
grpcio-status==1.48.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
jax==0.7.0
jaxlib==0.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kaggle==2.0.0
kagglesdk==0.1.15
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langsmith==0.1.147
llama-cloud==0.1.35
llama-cloud-services==0.6.53
llama-index==0.12.52
llama-index-agent-openai==0.4.12
llama-index-cli==0.4.4
llama-index-core==0.12.52.post1
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.8.0
llama-index-instrumentation==0.3.1
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.3
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.11
llama-index-readers-llama-parse==0.4.0
llama-index-workflows==1.2.0
llama-parse==0.6.53
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.3
mmh3==5.2.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
ollama==0.5.3
onnxruntime==1.23.1
openai==1.98.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
orjson==3.11.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
postgrest==2.28.2
posthog==5.4.0
primp==0.15.0
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyroaring==1.0.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
realtime==2.28.2
referencing==0.36.2
regex==2025.7.33
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.1
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
storage3==2.28.2
streamlit==1.47.1
StrEnum==0.4.15
strictyaml==1.7.3
striprtf==0.0.26
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
sympy==1.14.0
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.1
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.2
xgboost==3.1.3
yarl==1.23.0
zipp==3.23.0
zstandard==0.23.0
